renaming header to nav for consistency
